requirements.txt
appdirs==1.4.4
attrs==23.1.0
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
comet-ml==3.35.3
configobj==5.0.8
cryptography @ file:///croot/cryptography_1694444244250/work
docker-pycreds==0.4.0
dulwich==0.21.6
everett==3.1.0
filelock @ file:///croot/filelock_1700591183607/work
gitdb==4.0.11
GitPython==3.1.40
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
idna @ file:///croot/idna_1666125576474/work
importlib-resources==6.1.1
Jinja2 @ file:///croot/jinja2_1666908132255/work
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
markdown-it-py==3.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
Pillow @ file:///croot/pillow_1696580024257/work
pkgutil_resolve_name==1.3.10
protobuf==4.25.1
psutil==5.9.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.17.2
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-box==6.1.0
PyYAML @ file:///croot/pyyaml_1698096049011/work
referencing==0.31.1
requests @ file:///croot/requests_1690400202158/work
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.13.2
semantic-version==2.10.0
sentry-sdk==1.38.0
setproctitle==1.3.3
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sympy @ file:///croot/sympy_1701397643339/work
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
urllib3 @ file:///croot/urllib3_1698257533958/work
wandb==0.16.0
websocket-client==1.3.3
wrapt==1.16.0
wurlitzer==3.0.3
zipp==3.17.0
